Moves SelectOptionProps as common type
