+++ src/Components/Toolbar/Groups/FiltersCategoriesGroup.tsx
@@ -1,18 +1,9 @@
 import { ToolbarGroup } from '@patternfly/react-core/dist/dynamic/components/Toolbar';
 import React, { FunctionComponent, useState } from 'react';
 import { optionsForCategories } from '../constants';
-import { AttributeType, SetValues } from '../types';
+import { AttributeType, SelectOptionProps, SetValues } from '../types';
 import CategoryDropdown from './CategoryDropdown';
 import ToolbarInput from './ToolbarInput';
-
-// SelectOptionProps type for compatibility
-export type SelectOptionProps = {
-  key?: string;
-  value?: string;
-  description?: string;
-  isDisabled?: boolean;
-  [key: string]: any;
-};
 
 interface Props {
   filterCategories: Record<string, SelectOptionProps[]>;
