+++ src/Components/Toolbar/Groups/QuickDateGroup.tsx
@@ -4,17 +4,8 @@ import { SplitItem } from '@patternfly/react-core/dist/dynamic/layouts/Split';
 import { Split } from '@patternfly/react-core/dist/dynamic/layouts/Split';
 import React, { FunctionComponent } from 'react';
 import { today } from '../../../Utilities/helpers';
-import { AttributeType, SetValues } from '../types';
+import { AttributeType, SelectOptionProps, SetValues } from '../types';
 import ToolbarInput from './ToolbarInput';
-
-// SelectOptionProps type for compatibility
-export type SelectOptionProps = {
-  key?: string;
-  value?: string;
-  description?: string;
-  isDisabled?: boolean;
-  [key: string]: any;
-};
 
 const getDateByDays = (days: number): string =>
   today(days).toISOString().split(/T/)[0];
